feature/ranking

+++ src/main/kotlin/syntax/backend/runways/dto/UserRankingDTO.kt
@@ -4,5 +4,5 @@ data class UserRankingDTO (
     val id : String,
     val profileImage : String?,
     val nickname : String?,
-    val experience : Float
+    val experience : Int
 )
